Reduced key size from 64 to 32 bits

+++ src/main/scala/sword/langbook/db/interfaces.scala
@@ -96,9 +96,9 @@ trait RegisterDefinition {
 
 object Register {
   type Index = Int
-  type Key = Long
+  type Key = Int
   type Position = Int
-  type SetId = Long
+  type SetId = Int
   type UnicodeType = Int
 }
